add planet

+++ app/src/main/java/ca/cmpt276/cosmos/models/Spaceship.java
@@ -69,7 +69,7 @@ public class Spaceship {
         this.angle += r;
     }
 
-    public void incrementForwardVelocity(int increment) {
+    public void incrementForwardVelocity(double increment) {
         this.spaceshipForwardVel = spaceshipForwardVel + increment;
     }
 }
